+++ src/lightsclient/Part.java
@@ -11,8 +11,8 @@ public class Part {
 
 	private int channel;
 	private ArrayList<Event> notes;
-	private ArrayList<Float> partTimes;
-	private ArrayList<Float> measureTimes;
+	private ArrayList<Double> partTimes;
+	private ArrayList<Double> measureTimes;
 
 	private int possibleEvent = 0;
 	// private int possiblePhrase = 0;
@@ -40,11 +40,11 @@ public class Part {
 		return channel;
 	}
 
-	public float getTime() {
+	public double getTime() {
 		return notes.get(currentEvent).getTime();
 	}
 
-	public float getNextTime() {
+	public double getNextTime() {
 		if (currentEvent < notes.size()) {
 			return notes.get(currentEvent + 1).getTime();
 		} else {
@@ -52,22 +52,22 @@ public class Part {
 		}
 	}
 
-	public float getPartTime() {
+	public double getPartTime() {
 		return partTimes.get(currentPart);
 	}
 
-	public float getMeasureTime() {
+	public double getMeasureTime() {
 		return measureTimes.get(currentMeasure);
 	}
 
-	public void addMeasures(ArrayList<Float> measures) {
+	public void addMeasures(ArrayList<Double> measures) {
 		measureTimes = measures;
 
 		// populate measureIndexes
 		measureIndexes = new int[measureTimes.size()];
-		float measureTime = measureTimes.get(currentMeasure);
+		double measureTime = measureTimes.get(currentMeasure);
 		for (int i = 0; i < notes.size(); i++) {
-			float eventTime = notes.get(i).getTime();
+			double eventTime = notes.get(i).getTime();
 
 			// check if current time is past the measure marker
 			if (eventTime >= measureTime) {
@@ -112,14 +112,14 @@ public class Part {
 		currentMeasure = 0;
 	}
 
-	public void addParts(ArrayList<Float> parts) {
+	public void addParts(ArrayList<Double> parts) {
 		partTimes = parts;
 
 		// populate partIndexes
 		partIndexes = new int[partTimes.size()];
-		float partTime = partTimes.get(currentPart);
+		double partTime = partTimes.get(currentPart);
 		for (int i = 0; i < notes.size(); i++) {
-			float eventTime = notes.get(i).getTime();
+			double eventTime = notes.get(i).getTime();
 
 			if (eventTime >= partTime) {
 				partIndexes[currentPart] = i;
@@ -252,7 +252,7 @@ public class Part {
 		return correct;
 	}
 
-	public Float isPossible(int input) {
+	public Double isPossible(int input) {
 		// make sure that we aren't already looking at a possible new place
 		if (possibleEvent != currentEvent) {
 			Event ev = notes.get(possibleEvent);
@@ -287,7 +287,7 @@ public class Part {
 		}
 
 		// check next handful of notes
-		Float time = possibleEvent(input, currentEvent);
+		Double time = possibleEvent(input, currentEvent);
 		if (time != null) {
 			System.out.println("Possible new note on channel " + channel);
 			previousPossibleNotes.add(notes.get(possibleEvent));
@@ -328,7 +328,7 @@ public class Part {
 		return null;
 	}
 
-	private Float possibleEvent(int input, int index) {
+	private Double possibleEvent(int input, int index) {
 		// check next handful of events
 		for (int i = 1; i < 7; i++) {
 			if ((index + i) > notes.size()) {
@@ -446,7 +446,7 @@ public class Part {
 		}
 	}
 
-	public Float nextPart() {
+	public Double nextPart() {
 		// make sure we are not out of parts
 		if (currentPart == partTimes.size()) {
 			return null;
@@ -467,7 +467,7 @@ public class Part {
 		}
 
 		// find new currentNote and currentMeasure
-		float partTime = partTimes.get(currentPart);
+		double partTime = partTimes.get(currentPart);
 
 		// measures first
 		while (measureTimes.get(currentMeasure) < partTime) {
@@ -485,7 +485,7 @@ public class Part {
 		return partTimes.get(currentPart);
 	}
 
-	public Float nextMeasure() {
+	public Double nextMeasure() {
 		// make sure we are not out of measures
 		if (currentMeasure == measureTimes.size()) {
 			return null;
@@ -502,11 +502,11 @@ public class Part {
 		}
 
 		// find new currentNote and currentPart
-		float measureTime = measureTimes.get(currentMeasure);
+		double measureTime = measureTimes.get(currentMeasure);
 
 		// part first
 		// check if we even have to move
-		float nextPart = partTimes.get(currentPart + 1);
+		double nextPart = partTimes.get(currentPart + 1);
 		if (nextPart <= measureTime) {
 			while (partTimes.get(currentPart) < measureTime) {
 				currentPart++;
@@ -524,18 +524,18 @@ public class Part {
 		return measureTimes.get(currentMeasure);
 	}
 
-	public Float previousPart() {
+	public Double previousPart() {
 		// make sure we are not at the beginning of the song still
 		if (currentPart == 0) {
 			currentMeasure = 0;
 			currentEvent = 0;
-			return new Float(0);
+			return new Double(0);
 		}
 
 		currentPart--;
 
 		// find new currentNote and currentMeasure
-		float partTime = partTimes.get(currentPart);
+		double partTime = partTimes.get(currentPart);
 
 		// measures first
 		while (measureTimes.get(currentMeasure - 1) > partTime) {
@@ -550,18 +550,18 @@ public class Part {
 		return partTimes.get(currentPart);
 	}
 
-	public Float previousMeasure() {
+	public Double previousMeasure() {
 		// make sure we are not at the beginning of the song still
 		if (currentMeasure == 0) {
 			currentPart = 0;
 			currentEvent = 0;
-			return new Float(0);
+			return new Double(0);
 		}
 
 		currentMeasure--;
 
 		// find new currentNote and currentPart
-		float measureTime = measureTimes.get(currentMeasure);
+		double measureTime = measureTimes.get(currentMeasure);
 
 		// check if even necessary for part
 		if (partTimes.get(currentPart - 1) > measureTime) {
@@ -579,7 +579,7 @@ public class Part {
 		return measureTimes.get(currentMeasure);
 	}
 
-	public void changeTime(float newTime) {
+	public void changeTime(double newTime) {
 		if (newTime > getTime()) {
 			// if we are behind, find new note
 			while (getTime() < newTime) {
